Refactored is reproducing same tracks as develop branch
For AtMlesacMod, in the refactored class it is occationally finding an
additional track. These are physical tracks (looking at the event
display). The tracks they both find are identical so I think there
must be difference in the parameters I was using to process the two
runs but I can't find it anywhere.