Fist pass at extracting model information from AtRansacMod
Model information (shape) now moved to a class. Very dirty and needs
heacy rework. I don't know how I feel about the Model storing the hit
array. It should be data structure agnostic. It is only needed for the
sampling, so maybe that should live in AtRansac still, and the model
is constructed with just the AtHits required to fully define the
model. The model can then just be a container for the fitParameters
basically. Need to think about the relationship between model and
track. Right now I think model is purely shape information, track is
shape + additional info. So AtTrack will store an AtModel. Could also
simplify ananlysis as we can then ask the model directly for things
like position in 3D space.