Refactor AtTrackModel
Removed random sampling from AtTrackModel and begin a bit of the work
to generalize to arbitrary points. These methods should be moved
outside of AtRansacMod to a new base class.

Simplified AtTrackModel interface. It no longer cares about the data strucures
being used, instead works in terms of AtHits, XYZPoint, and
charge. Instead it simply stores the parameters to define the model, and
a parameter to describe how good the model is at describing the data.